refactor(pi): decompose subagent progress correlation
